revise: rename package name
