+++ spark/dl/pom.xml
@@ -81,6 +81,17 @@
             </exclusions>
         </dependency>
         <dependency>
+          <groupId>com.intel.analytics.bigdl.quantization</groupId>
+          <artifactId>quantization-java</artifactId>
+          <version>${project.version}</version>
+          <exclusions>
+            <exclusion>
+              <groupId>com.intel.analytics.bigdl.quantization</groupId>
+              <artifactId>quantization-native</artifactId>
+            </exclusion>
+          </exclusions>
+        </dependency>
+        <dependency>
             <groupId>org.apache.spark</groupId>
             <artifactId>spark-core_${scala.major.version}</artifactId>
             <version>${spark.version}</version>
@@ -311,36 +322,6 @@
             <properties>
                 <tagsToExclude>com.intel.analytics.bigdl.tags.Serial,com.intel.analytics.bigdl.tags.Parallel</tagsToExclude>
             </properties>
-        </profile>
-        <profile>
-            <id>with-fixpoint</id>
-            <dependencies>
-                <dependency>
-                    <groupId>com.intel.analytics.bigdl.fixpoint</groupId>
-                    <artifactId>fixpoint-java</artifactId>
-                    <version>${project.version}</version>
-                    <exclusions>
-                        <exclusion>
-                            <groupId>com.intel.analytics.bigdl.fixpoint</groupId>
-                            <artifactId>fixpoint-native</artifactId>
-                        </exclusion>
-                    </exclusions>
-                </dependency>
-            </dependencies>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>net.alchim31.maven</groupId>
-                        <artifactId>scala-maven-plugin</artifactId>
-                        <version>3.2.0</version>
-                        <configuration>
-                            <excludes>
-                                <exclude>com/intel/analytics/bigdl/example/udfpredictor/*Stream*.scala</exclude>
-                            </excludes>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
         </profile>
     </profiles>
 </project>
