+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/quantization/Linear.scala
@@ -14,18 +14,16 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.bigdl.nn.fixpoint
+package com.intel.analytics.bigdl.nn.quantization
 
+import com.intel.analytics.bigdl.nn.abstractnn.TensorModule
+import com.intel.analytics.bigdl.nn.{ErrorInfo, Module, Quantize}
+import com.intel.analytics.bigdl.quantization.Quantization
+import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
+import com.intel.analytics.bigdl.tensor.{FloatType, QuantizeTensor, Tensor}
+import com.intel.analytics.bigdl.utils.{T, Table}
 import java.nio.ByteBuffer
 import java.nio.file.{Files, Paths}
-
-import com.intel.analytics.bigdl.fixpoint.FixPoint
-import com.intel.analytics.bigdl.nn.{ErrorInfo, Module, Quantize}
-import com.intel.analytics.bigdl.nn.abstractnn.TensorModule
-import com.intel.analytics.bigdl.tensor.{FloatType, QuantizeTensor, Tensor}
-import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
-import com.intel.analytics.bigdl.utils.{T, Table}
-
 import scala.reflect.ClassTag
 
 class Linear[T: ClassTag](
@@ -87,15 +85,15 @@ class Linear[T: ClassTag](
   def init(): this.type = {
     val byteArrayOfWeight = weight.getStorage.get
     weight.setStorageInJni(
-      FixPoint.FixFCKernelDescInit(outputSize, inputSize))
+      Quantization.FixFCKernelDescInit(outputSize, inputSize))
 
     ev.getType() match {
       case FloatType =>
         val minArray = min.asInstanceOf[Array[Float]]
         val maxArray = max.asInstanceOf[Array[Float]]
 
-        FixPoint.FixFCKernelLoadFromModel(weight.getStorageInJni, byteArrayOfWeight,
-          minArray, maxArray, outputSize, inputSize, WEIGHT_THRESHOLD, FixPoint.NCHW)
+        Quantization.FixFCKernelLoadFromModel(weight.getStorageInJni, byteArrayOfWeight,
+          minArray, maxArray, outputSize, inputSize, WEIGHT_THRESHOLD, Quantization.NCHW)
       case _ => throw new UnsupportedOperationException(s"Only support Float for quantized model")
     }
 
@@ -121,7 +119,7 @@ class Linear[T: ClassTag](
     }
 
     if (!data.isInitialized) {
-      data.setStorageInJni(FixPoint.FixFCDataDescInit(batchSize, inputSize))
+      data.setStorageInJni(Quantization.FixFCDataDescInit(batchSize, inputSize))
     }
 
     ev.getType() match {
@@ -129,8 +127,8 @@ class Linear[T: ClassTag](
         val src = input.storage().array().asInstanceOf[Array[Float]]
         val offset = input.storageOffset() - 1
 
-        FixPoint.FixFCDataInit(data.getStorageInJni, src, offset, batchSize, inputSize,
-          THRESHOLD, FixPoint.NCHW)
+        Quantization.FixFCDataInit(data.getStorageInJni, src, offset, batchSize, inputSize,
+          THRESHOLD, Quantization.NCHW)
 
         val outputArray = output.storage().array().asInstanceOf[Array[Float]]
         val outputOffset = output.storageOffset() - 1
@@ -139,8 +137,8 @@ class Linear[T: ClassTag](
         val biasArray = bias.storage().array().asInstanceOf[Array[Float]]
         val biasOffset = bias.storageOffset() - 1
 
-        FixPoint.InternalMixPrecisionConvolutionGEMM(
-          FixPoint.NCHW, weight.getStorageInJni, data.getStorageInJni, outputArray, outputOffset,
+        Quantization.InternalMixPrecisionConvolutionGEMM(
+          Quantization.NCHW, weight.getStorageInJni, data.getStorageInJni, outputArray, outputOffset,
           weightSumArray, weightSumOffset, biasArray, biasOffset,
           batchSize, outputSize, 1, 1,
           FAULT_TOLERANCE)
