+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/quantization/LinearSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.bigdl.nn.fixpoint
+package com.intel.analytics.bigdl.nn.quantization
 
 import com.intel.analytics.bigdl.nn.{Linear => NNLinear}
 import com.intel.analytics.bigdl.numeric.NumericFloat
