+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/quantization/SpatialConvolutionSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.bigdl.nn.fixpoint
+package com.intel.analytics.bigdl.nn.quantization
 
 import com.intel.analytics.bigdl.nn.{SpatialConvolution => NNSpatialConvolution}
 import com.intel.analytics.bigdl.numeric.NumericFloat
